MDEV-34705: Binlog-in-engine: Refactor InnoDB part
Move high-level binlog code to handler/handler0binlog.cc and low-level code
to fsp/fsp0binlog.cc.
