+++ storage/innobase/handler/ha_innodb.cc
@@ -108,6 +108,7 @@ extern my_bool opt_readonly;
 #include "fil0pagecompress.h"
 #include "ut0mem.h"
 #include "row0ext.h"
+#include "handler0binlog.h"
 #include "mariadb_stats.h"
 simple_thread_local ha_handler_stats *mariadb_stats;
 
@@ -4200,8 +4201,8 @@ static int innodb_init(void* p)
         innobase_hton->update_optimizer_costs= innobase_update_optimizer_costs;
         innobase_hton->binlog_init= innodb_binlog_init;
         innobase_hton->binlog_write_direct= innobase_binlog_write_direct;
-        innobase_hton->binlog_oob_data= fsp_binlog_oob;
-        innobase_hton->binlog_oob_free= fsp_free_oob;
+        innobase_hton->binlog_oob_data= innodb_binlog_oob;
+        innobase_hton->binlog_oob_free= innodb_free_oob;
         innobase_hton->get_binlog_reader= innodb_get_binlog_reader;
 
 	innodb_remember_check_sysvar_funcs();
