+++ storage/innobase/include/ut0new.h
@@ -849,9 +849,9 @@ constexpr const char* const auto_event_names[] =
   "dict0stats",
   "eval0eval",
   "fil0crypt",
+  "fsp0binlog",
   "fil0fil",
   "fsp0file",
-  "fsp0fsp",
   "fts0ast",
   "fts0blex",
   "fts0config",
@@ -865,6 +865,7 @@ constexpr const char* const auto_event_names[] =
   "gis0sea",
   "ha_innodb",
   "handler0alter",
+  "handler0binlog",
   "hash0hash",
   "i_s",
   "lexyy",
