+++ storage/innobase/trx/trx0trx.cc
@@ -47,6 +47,7 @@ Created 3/26/1996 Heikki Tuuri
 #include "trx0xa.h"
 #include "ut0pool.h"
 #include "ut0vec.h"
+#include "handler0binlog.h"
 #include "log.h"
 
 #include <set>
@@ -1170,7 +1171,7 @@ inline void trx_t::write_serialisation_history(mtr_t *mtr)
       trx_sys.assign_new_trx_no(this);
 
     /* Include binlog data in the commit record, if any. */
-    fsp_binlog_trx(this, mtr);
+    innodb_binlog_trx(this, mtr);
 
     UT_LIST_REMOVE(rseg->undo_list, undo);
     /* Change the undo log segment state from TRX_UNDO_ACTIVE, to
